+++ app/models/course_model.py
@@ -11,11 +11,11 @@ class Courses(db.Base):
 
     pcd = db.Column(db.Boolean)
 
-    employability_index = db.Column(db.Integer, nullable=True)
+    employability_index = db.Column(db.Float, nullable=True)
 
-    businessperson_index = db.Column(db.Integer, nullable=True)
+    businessperson_index = db.Column(db.Float, nullable=True)
 
-    public_server_index = db.Column(db.Integer, nullable=True)
+    public_server_index = db.Column(db.Float, nullable=True)
 
     # tmax 255, link do site da empresa (se tiver)
     link_site = db.Column(db.String(255), nullable=True)
